aws: restructure files
