+++ packages/aws/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "rpaframework-google"
-version = "1.0.0"
+name = "rpaframework-aws"
+version = "0.2.0"
 description = "AWS library for RPA Framework"
 authors = [
 	"RPA Framework <rpafw@robocorp.com>",
@@ -44,7 +44,7 @@ amazon-textract-response-parser = { version = "^0.1.1", optional = true }
 [tool.poetry.dev-dependencies]
 black = {version = "^21.5b0", allow-prereleases = true}
 flake8 = "^3.7.9"
-pylint = "^2.4.4, <2.13.0"
+pylint = "^2.4.4, <2.13"
 pytest = "^6.1.2"
 mock = "^4.0.2"
 pytest-cov = "^2.10.1"
