+++ packages/aws/tests/init_environment_variables.sh
@@ -1,13 +1,3 @@
-export GOOGLE_TEST_USE_VAULT=FALSE
-export GOOGLE_TEST_PROJECT_ID=<PROJECTID>
-export GOOGLE_TEST_SHEET=<SHEETID>
-export GOOGLE_TEST_SERVICE_ACCOUNT=/Users/<USER>/service_account.json
-export GOOGLE_TEST_SCRIPT_ID=<SCRIPTID>
-export GOOGLE_TEST_STORAGE_BUCKET=<BUCKET_NAME>
-export GOOGLE_TEST_VAULT_NAME=<VAULT_NAME>
-export GOOGLE_TEST_VAULT_KEY=<VAULT_KEY>
-# token or serviceaccount
-export GOOGLE_TEST_AUTH_TYPE=<TYPE>
-export RPA_SECRET_MANAGER=RPA.Robocorp.Vault.FileSecrets
-export RPA_SECRET_FILE=/Users/<USER>/secrets.json
-export GOOGLE_TEST_RESOURCE_DIR=/Users/<USER>/testdata
+export AWS_KEY_ID=<AWS_KEY>
+export AWS_KEY_SECRET=<AWS_KEY_SECRET>
+export AWS_REGION=us-east-1
