+++ packages/main/pyproject.toml
@@ -38,6 +38,7 @@ packages = [
 python = "^3.6.2"
 docutils = "*"
 dataclasses = { version = "^0.7", python=">=3.6,<3.7" }
+rpaframework-aws = { version = "^0.1.0", optional = true }
 rpaframework-core = "^7.0.1"
 rpaframework-dialogs = "^1.0.0"
 rpaframework-pdf = "^2.0.0"
@@ -106,6 +107,7 @@ rpa-crypto = 'RPA.scripts.crypto:main'
 use-robocorp-vault = 'RPA.scripts.robocorp_cloud:main'
 
 [tool.poetry.extras]
+aws = ["rpaframework-aws"]
 cv = ["rpaframework-recognition"]
 playwright = ["robotframework-browser"]
